#49 Adding capability to put and get tags - Code Review changes
