+++ server/src/main/java/com/adobe/testing/s3mock/FileStoreController.java
@@ -31,8 +31,28 @@ import static org.springframework.http.HttpStatus.OK;
 import static org.springframework.http.HttpStatus.PARTIAL_CONTENT;
 import static org.springframework.http.HttpStatus.REQUESTED_RANGE_NOT_SATISFIABLE;
 
-import com.adobe.testing.s3mock.domain.*;
-import com.adobe.testing.s3mock.dto.*;
+import com.adobe.testing.s3mock.domain.Bucket;
+import com.adobe.testing.s3mock.domain.BucketContents;
+import com.adobe.testing.s3mock.domain.FileStore;
+import com.adobe.testing.s3mock.domain.S3Exception;
+import com.adobe.testing.s3mock.domain.S3Object;
+import com.adobe.testing.s3mock.domain.Tag;
+import com.adobe.testing.s3mock.dto.BatchDeleteRequest;
+import com.adobe.testing.s3mock.dto.BatchDeleteResponse;
+import com.adobe.testing.s3mock.dto.CompleteMultipartUploadResult;
+import com.adobe.testing.s3mock.dto.CopyObjectResult;
+import com.adobe.testing.s3mock.dto.CopyPartResult;
+import com.adobe.testing.s3mock.dto.DeletedObject;
+import com.adobe.testing.s3mock.dto.InitiateMultipartUploadResult;
+import com.adobe.testing.s3mock.dto.ListAllMyBucketsResult;
+import com.adobe.testing.s3mock.dto.ListBucketResult;
+import com.adobe.testing.s3mock.dto.ListMultipartUploadsResult;
+import com.adobe.testing.s3mock.dto.ListPartsResult;
+import com.adobe.testing.s3mock.dto.MultipartUpload;
+import com.adobe.testing.s3mock.dto.ObjectRef;
+import com.adobe.testing.s3mock.dto.Owner;
+import com.adobe.testing.s3mock.dto.Range;
+import com.adobe.testing.s3mock.dto.Tagging;
 
 import java.io.FileInputStream;
 import java.io.IOException;
