+++ server/src/main/java/com/adobe/testing/s3mock/S3MockApplication.java
@@ -23,7 +23,19 @@ import com.adobe.testing.s3mock.domain.Bucket;
 import com.adobe.testing.s3mock.domain.FileStore;
 import com.adobe.testing.s3mock.domain.KMSKeyStore;
 import com.adobe.testing.s3mock.domain.Tag;
-import com.adobe.testing.s3mock.dto.*;
+import com.adobe.testing.s3mock.dto.BatchDeleteRequest;	
+import com.adobe.testing.s3mock.dto.BatchDeleteResponse;
+import com.adobe.testing.s3mock.dto.CompleteMultipartUploadResult;	
+import com.adobe.testing.s3mock.dto.CopyObjectResult;	
+import com.adobe.testing.s3mock.dto.CopyPartResult;	
+import com.adobe.testing.s3mock.dto.ErrorResponse;	
+import com.adobe.testing.s3mock.dto.InitiateMultipartUploadResult;	
+import com.adobe.testing.s3mock.dto.ListAllMyBucketsResult;	
+import com.adobe.testing.s3mock.dto.ListBucketResult;	
+import com.adobe.testing.s3mock.dto.ListMultipartUploadsResult;	
+import com.adobe.testing.s3mock.dto.ListPartsResult;	
+import com.adobe.testing.s3mock.dto.Owner;
+import com.adobe.testing.s3mock.dto.Tagging;
 import com.adobe.testing.s3mock.util.ObjectRefConverter;
 import com.adobe.testing.s3mock.util.RangeConverter;
 import com.thoughtworks.xstream.XStream;
