+++ testsupport/testng/src/test/java/com/adobe/testing/s3mock/testng/S3MockListenerXMLConfigurationTest.java
@@ -18,15 +18,14 @@ package com.adobe.testing.s3mock.testng;
 
 import com.adobe.testing.s3mock.util.HashUtil;
 import com.amazonaws.services.s3.AmazonS3;
-import com.amazonaws.services.s3.model.*;
+import com.amazonaws.services.s3.model.PutObjectRequest;
+import com.amazonaws.services.s3.model.S3Object;
 import org.testng.Assert;
 import org.testng.annotations.Test;
 
 import java.io.File;
 import java.io.FileInputStream;
 import java.io.InputStream;
-import java.util.ArrayList;
-import java.util.List;
 
 @Test
 public class S3MockListenerXMLConfigurationTest {
@@ -54,37 +53,5 @@ public class S3MockListenerXMLConfigurationTest {
         s3Object.close();
 
         Assert.assertEquals(uploadHash, downloadedHash, "Up- and downloaded Files should have equal Hashes");
-    }
-
-    /**
-     * Creates a bucket, stores a file, adds tags, retrieves tags and checks them for consistency.
-     */
-    @Test
-    public void shouldAddAndRetrieveTags() {
-        final File uploadFile = new File(UPLOAD_FILE_NAME);
-
-        s3Client.createBucket(BUCKET_NAME);
-        s3Client.putObject(new PutObjectRequest(BUCKET_NAME, uploadFile.getName(), uploadFile));
-
-        final S3Object s3Object = s3Client.getObject(BUCKET_NAME, uploadFile.getName());
-
-        GetObjectTaggingRequest getObjectTaggingRequest = new GetObjectTaggingRequest(BUCKET_NAME,s3Object.getKey());
-        GetObjectTaggingResult getObjectTaggingResult = s3Client.getObjectTagging(getObjectTaggingRequest);
-
-        // There shouldn't be any tags here
-        Assert.assertEquals(getObjectTaggingResult.getTagSet().size(),0);
-
-        List<Tag> tagList = new ArrayList<>();
-        tagList.add(new Tag("foo","bar"));
-
-        final SetObjectTaggingRequest setObjectTaggingRequest = new SetObjectTaggingRequest(BUCKET_NAME,s3Object.getKey(),new ObjectTagging(tagList));
-        final SetObjectTaggingResult setObjectTaggingResult = s3Client.setObjectTagging(setObjectTaggingRequest);
-
-        getObjectTaggingRequest = new GetObjectTaggingRequest(BUCKET_NAME,s3Object.getKey());
-        getObjectTaggingResult = s3Client.getObjectTagging(getObjectTaggingRequest);
-
-        // There should be 'foo:bar' here
-        Assert.assertEquals(getObjectTaggingResult.getTagSet().size(),1);
-        Assert.assertEquals(getObjectTaggingResult.getTagSet().get(0).getValue(),"bar");
     }
 }
