+++ testsupport/testng/src/test/resources/testng.xml
@@ -27,7 +27,6 @@
           <class name="com.adobe.testing.s3mock.testng.S3MockListenerXMLConfigurationTest">
               <methods>
                   <include name="shouldUploadAndDownloadObject"/>
-                  <include name="shouldAddAndRetrieveTags"/>
               </methods>
           </class>
       </classes>
